admin/backend/requirements.txt
# Super Admin Dashboard backend dependencies
#
# For Developers:
#   Install with: pip install -r requirements.txt
#   Core: FastAPI + SQLAlchemy async + PostgreSQL
#   Auth: bcrypt for password hashing, PyJWT for tokens
#   HTTP: httpx for proxying to the LLM Gateway and health checks
#
# For QA Engineers:
#   Test dependencies (pytest, pytest-asyncio) are at the bottom.
#   Run tests with: pytest -v

# Web framework
fastapi>=0.110.0
uvicorn>=0.27.0

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0

# Settings & validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Authentication
bcrypt>=4.0.0
PyJWT>=2.8.0

# HTTP client (proxy & health checks)
httpx>=0.27.0

# Multipart form support
python-multipart>=0.0.6

# Dev / test
pytest>=8.0.0
pytest-asyncio>=0.23.0
